requirements.txt
httpx
tqdm
python-dotenv